Updated SCInput()
Testing out ignoring punctuation in the SCInput function. More work needs to be done though because it doesn't put the punctuation back at the moment, it just returns the spellchecked word (ignoring punctuation). It's also a test, which is why I created a new branch for it, since it isn't fully tested (tested it on an online IDE on a Chromebook so can't guarantee it's working). 
So keep all changes to this file to this branch for now.